// api/endpoints/tixTrack.schemas.ts
export interface SubjectResponseDto {
  /** Id of the subject */
  id: string;
  /**
   * Name of the subject
   * @minLength 1
   * @maxLength 255
   */
  name: string;
  /**
   * Subject description
   * @minLength 1
   * @maxLength 255
   */
  description?: string;
  /** Service ID of the subject */
  serviceId: string;
}